refactor(transformers): restructure transformer blocks/ops namespaces.
