+++ crates/bunsen/src/blocks/transformers/attention/mod.rs
@@ -1,21 +1,4 @@
-//! # Attention Extensions
-
-mod csa;
-mod kv_attention;
-mod kvcache;
-mod masks;
-mod multihead_utils;
-mod sdpa;
-
-#[doc(inline)]
-pub use csa::*;
-#[doc(inline)]
-pub use kv_attention::*;
-#[doc(inline)]
-pub use kvcache::*;
-#[doc(inline)]
-pub use masks::*;
-#[doc(inline)]
-pub use multihead_utils::*;
-#[doc(inline)]
-pub use sdpa::*;
+//! Attention Blocks.
+pub mod csa;
+pub mod kvcache;
+pub mod sdpa;
