+++ crates/bunsen/src/kits/gpts/nanochat/blocks/block.rs
@@ -15,10 +15,12 @@ use burn::{
 use crate::{
     blocks::transformers::{
         attention::{
-            CausalSelfAttention,
-            CausalSelfAttentionConfig,
-            CausalSelfAttentionMeta,
-            KVCache,
+            csa::{
+                CausalSelfAttention,
+                CausalSelfAttentionConfig,
+                CausalSelfAttentionMeta,
+            },
+            kvcache::KVCache,
         },
         embedding::RotaryEmbedding,
         mlp::{
