+++ crates/bunsen/src/ops/transformers/attention/attend.rs
@@ -1,55 +1,14 @@
-//! # Key/value cached attention over `MultiHeadAttention` weights.
-//!
-//! A genuine KV cache: keys and values are kept **projected and head-split**,
-//! so a decode step attends over the whole history while projecting only the
-//! new tokens, and the attention itself only scores the new queries.
-//!
-//! This reuses [`MultiHeadAttention`]'s own `Linear` layers rather than
-//! replacing the module, so weights, loading and any cross-checks stay exactly
-//! as they are — only the forward path differs.
-//!
-//! ## Why not `MhaCache`
-//!
-//! [`MhaCache`](burn::nn::attention::MhaCache) caches the *projections* and
-//! expects the whole sequence back on every call, so its attention still scores
-//! the full prefix each step. That is a real saving for cross-attention, whose
-//! keys and values never change, but it leaves self-attention quadratic. These
-//! functions take only the new tokens.
-//!
-//! ## Why not `KVCache`
-//!
-//! [`KVCache`](super::KVCache) is a preallocated *multi-layer* cache: a single
-//! `Tensor<B, 6>` over `[layers, kv, batch, heads, seq, d_k]`, grown in chunks,
-//! and a `Module` in its own right. It wants the geometry — layer count, head
-//! count, head dimension, batch size, and a sequence bound — declared up front,
-//! and it serves bunsen's own attention stack
-//! ([`CausalSelfAttention`](super::CausalSelfAttention)).
-//!
-//! [`AttnKv`] is a plain per-layer value over burn's [`MultiHeadAttention`]
-//! weights. It declares no geometry, and it also covers a case `KVCache` does
-//! not model: cross-attention keys and values, projected once from the encoder
-//! output and then reused unchanged rather than grown.
-//!
-//! Reach for `KVCache` when the geometry is known up front and the allocation
-//! matters; reach for these when the attention is burn's own and the cache is
-//! per-layer.
-//!
-//! ## Matching the uncached path
-//!
-//! The arithmetic mirrors `MultiHeadAttention::forward` exactly — the same
-//! `1/sqrt(d_k)` scaling, the same `min_float` mask fill, the same
-//! `quiet_softmax` switch — so a cached decode reproduces an uncached one.
-//! `layer_norm_self_attn_kv`'s contract test is what holds that true.
-//!
-//! Note this path deliberately omits the attention-score dropout that
-//! `MultiHeadAttention::attn_scores` applies: caching is an inference concern,
-//! and burn's `Dropout` is a no-op outside training anyway.
+//! Multihead Layer Normalized Attention Utilities.
 
 use burn::{
     Tensor,
     nn::{
         LayerNorm,
-        attention::MultiHeadAttention,
+        attention::{
+            MhaInput,
+            MhaOutput,
+            MultiHeadAttention,
+        },
     },
     prelude::{
         Backend,
@@ -61,57 +20,17 @@ use burn::{
     },
 };
 
-/// Projected, head-split keys and values: `[batch, heads, seq, d_k]`.
-///
-/// Built by [`project_kv`]. Self-attention grows one of these a step at a
-/// time; cross-attention builds one per layer and reuses it untouched.
-#[derive(Clone, Debug)]
-pub struct AttnKv<B: Backend> {
-    /// `[batch, heads, seq, d_k]` keys.
-    pub key: Tensor<B, 4>,
-
-    /// `[batch, heads, seq, d_k]` values.
-    pub value: Tensor<B, 4>,
-}
-
-impl<B: Backend> AttnKv<B> {
-    /// The number of cached positions.
-    pub fn seq_len(&self) -> usize {
-        self.key.dims()[2]
-    }
-
-    /// The batch size.
-    pub fn batch_size(&self) -> usize {
-        self.key.dims()[0]
-    }
-
-    /// Appends `next` along the sequence axis.
-    ///
-    /// This reallocates: every call copies the whole cache, so growing a
-    /// self-attention cache one token at a time costs O(T^2) copying across a
-    /// decode of length T. That is not worth avoiding at a few hundred
-    /// positions, but a long-context decode wants a preallocated cache
-    /// instead; see [`KVCache`](super::KVCache).
-    pub fn concat(
-        self,
-        next: Self,
-    ) -> Self {
-        Self {
-            key: Tensor::cat(vec![self.key, next.key], 2),
-            value: Tensor::cat(vec![self.value, next.value], 2),
-        }
-    }
-}
+use super::AttnKvPair;
 
 /// Projects `[batch, seq, d_model]` into head-split keys and values.
 ///
 /// For cross-attention this is the whole win: call it once per layer against
 /// the encoder output and the result serves every decode step.
-pub fn project_kv<B: Backend>(
+pub fn project_kv_pair<B: Backend>(
     mha: &MultiHeadAttention<B>,
     x: Tensor<B, 3>,
-) -> AttnKv<B> {
-    AttnKv {
+) -> AttnKvPair<B> {
+    AttnKvPair {
         key: split_heads(mha, mha.key.forward(x.clone())),
         value: split_heads(mha, mha.value.forward(x)),
     }
@@ -120,13 +39,116 @@ pub fn project_kv<B: Backend>(
 /// `[batch, seq, d_model]` -> `[batch, heads, seq, d_k]`.
 ///
 /// Mirrors `MultiHeadAttention::attention_linear`, minus the projection.
-fn split_heads<B: Backend>(
+pub fn split_heads<B: Backend>(
     mha: &MultiHeadAttention<B>,
     x: Tensor<B, 3>,
 ) -> Tensor<B, 4> {
     let [batch, seq, _] = x.dims();
     x.reshape([batch, seq, mha.n_heads, mha.d_k])
         .swap_dims(1, 2)
+}
+/// Computes layer normalized self-attn.
+///
+/// # Arguments
+/// * `layer_norm` - `LayerNorm`.
+/// * `mh_attn` - `MultiHeadAttention`.
+/// * `x` - `[batch, seq_len, d_model]` input.
+/// * `mask` - Optional `[batch, seq_len, seq_len]` attention mask.
+///
+/// # Returns
+/// `RdabForwardRecord` - forward record.
+/// * `fr.output` : `[batch, seq_len, d_model]`.
+/// * `fr.ca_weights` : `[batch, n_heads, seq_len, seq_len]`.
+pub fn layer_norm_self_attn<B: Backend>(
+    layer_norm: &LayerNorm<B>,
+    mh_attn: &MultiHeadAttention<B>,
+    x: Tensor<B, 3>,
+    mask: Option<Tensor<B, 3, Bool>>,
+) -> MhaOutput<B> {
+    #[cfg(any(debug_assertions, test))]
+    {
+        use crate::contracts::*;
+        let d_model = mh_attn.d_model;
+        assert_eq!(
+            d_model,
+            layer_norm.gamma.dims()[0],
+            "layer_norm dims ({}) != d_model ({d_model})",
+            layer_norm.gamma.dims()[0],
+        );
+
+        let [batch, seq_len] = unpack_shape_contract!(
+            ["batch", "seq_len", "d_model"],
+            &x,
+            &["batch", "seq_len"],
+            &[("d_model", d_model)]
+        );
+
+        if let Some(mask) = &mask {
+            let [mask_batch] = unpack_shape_contract!(
+                ["mask_batch", "seq_len", "seq_len"],
+                mask,
+                &["mask_batch"],
+                &[("seq_len", seq_len)]
+            );
+            if mask_batch != 1 {
+                assert_eq!(
+                    mask_batch, batch,
+                    "batch sizes not broadcastable {batch} vs {mask_batch}"
+                );
+            }
+        }
+    }
+
+    let input = MhaInput::self_attn(layer_norm.forward(x));
+    let input = match mask {
+        Some(mask) => input.mask_attn(mask),
+        None => input,
+    };
+    mh_attn.forward(input)
+}
+
+/// Computes layer normalized cross-attn.
+///
+/// # Arguments
+/// * `layer_norm` - `LayerNorm`.
+/// * `mh_attn` - `MultiHeadAttention`.
+/// * `x` - `[batch, seq_len, d_model]` input.
+/// * `xa` - `[batch, cross_len, d_model]` cross-attention input.
+///
+/// `cross_len` is independent of `seq_len`: cross-attention exists to attend a
+/// query sequence over a *different* one.
+///
+/// # Returns
+/// `RdabForwardRecord` - forward record.
+/// * `fr.output` : `[batch, seq_len, d_model]`.
+/// * `fr.ca_weights` : `[batch, n_heads, seq_len, cross_len]`.
+pub fn layer_norm_cross_attn<B: Backend>(
+    layer_norm: &LayerNorm<B>,
+    mh_attn: &MultiHeadAttention<B>,
+    x: Tensor<B, 3>,
+    xa: Tensor<B, 3>,
+) -> MhaOutput<B> {
+    #[cfg(any(debug_assertions, test))]
+    {
+        crate::contracts::define_shape_contract!(CONTRACT, ["batch", "seq_len", "d_model"]);
+        let d_model = mh_attn.d_model;
+        assert_eq!(
+            d_model,
+            layer_norm.gamma.dims()[0],
+            "layer_norm dims ({}) != d_model ({d_model})",
+            layer_norm.gamma.dims()[0],
+        );
+
+        let [batch, _seq_len] =
+            CONTRACT.unpack_shape(&x, &["batch", "seq_len"], &[("d_model", d_model)]);
+
+        // `xa` shares the batch and the model width, but NOT the sequence
+        // length — binding `seq_len` here would reject every real cross
+        // attention.
+        CONTRACT.assert_shape(&xa, &[("batch", batch), ("d_model", d_model)]);
+    };
+
+    mh_attn.forward(MhaInput::new(layer_norm.forward(x.clone()), xa.clone(), xa))
 }
 
 /// Scaled dot-product attention over head-split tensors, then the output
@@ -139,10 +161,10 @@ fn split_heads<B: Backend>(
 ///
 /// # Returns
 /// `[batch, seq_new, d_model]`.
-fn attend<B: Backend>(
+pub fn attend_q_kv_mask<B: Backend>(
     mha: &MultiHeadAttention<B>,
     q: Tensor<B, 4>,
-    kv: &AttnKv<B>,
+    kv: &AttnKvPair<B>,
     mask: Option<Tensor<B, 3, Bool>>,
 ) -> Tensor<B, 3> {
     let [batch, _, seq_new, _] = q.dims();
@@ -182,18 +204,18 @@ fn attend<B: Backend>(
 /// # Arguments
 /// * `x`: `[batch, seq_new, d_model]` — the new token(s) only.
 /// * `mask`: optional `[batch, seq_new, seq_past + seq_new]`. Build it with
-///   [`causal_mask(seq_new, seq_past, device)`](super::causal_mask). A single
+///   [`causal_mask(seq_new, seq_past, device)`](`super::causal_mask`). A single
 ///   new token needs no mask: it may attend to everything already cached.
 /// * `cache`: `None` at the start of a stream; grown in place thereafter.
 ///
 /// # Returns
 /// `[batch, seq_new, d_model]`.
-pub fn layer_norm_self_attn_kv<B: Backend>(
+pub fn layer_norm_self_attn_w_kv_cache<B: Backend>(
     layer_norm: &LayerNorm<B>,
     mha: &MultiHeadAttention<B>,
     x: Tensor<B, 3>,
     mask: Option<Tensor<B, 3, Bool>>,
-    cache: &mut Option<AttnKv<B>>,
+    cache: &mut Option<AttnKvPair<B>>,
 ) -> Tensor<B, 3> {
     #[cfg(any(debug_assertions, test))]
     crate::contracts::assert_shape_contract!(
@@ -205,14 +227,14 @@ pub fn layer_norm_self_attn_kv<B: Backend>(
     let normed = layer_norm.forward(x);
 
     let q = split_heads(mha, mha.query.forward(normed.clone()));
-    let fresh = project_kv(mha, normed);
+    let fresh = project_kv_pair(mha, normed);
 
     let grown = match cache.take() {
         Some(past) => past.concat(fresh),
         None => fresh,
     };
 
-    let out = attend(mha, q, &grown, mask);
+    let out = attend_q_kv_mask(mha, q, &grown, mask);
     *cache = Some(grown);
 
     out
@@ -221,20 +243,21 @@ pub fn layer_norm_self_attn_kv<B: Backend>(
 /// Layer-normed cross-attention against a fixed key/value cache.
 ///
 /// The keys and values come from the encoder output and never change, so
-/// [`project_kv`] runs once per layer per decode rather than once per token.
-/// No mask: every query may attend to the whole encoded sequence.
+/// [`project_kv_pair`] runs once per layer per decode rather
+/// than once per token. No mask: every query may attend to the whole encoded
+/// sequence.
 ///
 /// # Arguments
 /// * `x`: `[batch, seq_new, d_model]` — the new token(s) only.
-/// * `kv`: from [`project_kv`] over the encoder output.
+/// * `kv`: from [`project_kv_pair`] over the encoder output.
 ///
 /// # Returns
 /// `[batch, seq_new, d_model]`.
-pub fn layer_norm_cross_attn_kv<B: Backend>(
+pub fn layer_norm_cross_attn_w_kv_cache<B: Backend>(
     layer_norm: &LayerNorm<B>,
     mha: &MultiHeadAttention<B>,
     x: Tensor<B, 3>,
-    kv: &AttnKv<B>,
+    kv: &AttnKvPair<B>,
 ) -> Tensor<B, 3> {
     #[cfg(any(debug_assertions, test))]
     {
@@ -253,7 +276,7 @@ pub fn layer_norm_cross_attn_kv<B: Backend>(
     }
 
     let q = split_heads(mha, mha.query.forward(layer_norm.forward(x)));
-    attend(mha, q, kv, None)
+    attend_q_kv_mask(mha, q, kv, None)
 }
 
 #[cfg(test)]
@@ -273,15 +296,75 @@ mod tests {
         },
     };
 
-    use super::*;
+    use super::{
+        super::causal_mask,
+        *,
+    };
     use crate::{
-        blocks::transformers::attention::causal_mask,
         burner::tensor::TensorElemOpExt,
         support::testing::CpuBackend,
     };
 
     type B = CpuBackend;
     type F = <B as BackendTypes>::FloatElem;
+
+    /// Cross-attention must accept a `xa` whose sequence length differs from
+    /// the query's — that is the entire point of it.
+    ///
+    /// A contract that bound both to one `seq_len` would pass every
+    /// same-length unit test and still be wrong, which is why this one uses
+    /// deliberately mismatched lengths.
+    #[test]
+    fn test_cross_attn_accepts_a_different_cross_length() {
+        let device = Default::default();
+        let (batch, d_model, n_heads) = (2, 32, 4);
+        let (seq_len, cross_len) = (3, 17);
+
+        let attn = MultiHeadAttentionConfig::new(d_model, n_heads).init::<B>(&device);
+        let ln = LayerNormConfig::new(d_model).init::<B>(&device);
+
+        let x = Tensor::<B, 3>::random([batch, seq_len, d_model], Distribution::Default, &device);
+        let xa =
+            Tensor::<B, 3>::random([batch, cross_len, d_model], Distribution::Default, &device);
+
+        let out = layer_norm_cross_attn(&ln, &attn, x, xa);
+
+        // The output follows the query, and the weights span query x cross.
+        assert_eq!(out.context.dims(), [batch, seq_len, d_model]);
+        assert_eq!(out.weights.dims(), [batch, n_heads, seq_len, cross_len]);
+    }
+
+    /// The equal-length case still works, so the loosened contract did not
+    /// simply stop checking.
+    #[test]
+    fn test_cross_attn_accepts_equal_lengths() {
+        let device = Default::default();
+        let (batch, d_model, n_heads, seq_len) = (1, 16, 2, 5);
+
+        let attn = MultiHeadAttentionConfig::new(d_model, n_heads).init::<B>(&device);
+        let ln = LayerNormConfig::new(d_model).init::<B>(&device);
+
+        let x = Tensor::<B, 3>::random([batch, seq_len, d_model], Distribution::Default, &device);
+
+        let out = layer_norm_cross_attn(&ln, &attn, x.clone(), x);
+        assert_eq!(out.context.dims(), [batch, seq_len, d_model]);
+    }
+
+    /// A `d_model` mismatch is still rejected.
+    #[test]
+    #[should_panic(expected = "Shape Error")]
+    fn test_cross_attn_rejects_a_d_model_mismatch() {
+        let device = Default::default();
+        let (batch, d_model, n_heads) = (1, 16, 2);
+
+        let attn = MultiHeadAttentionConfig::new(d_model, n_heads).init::<B>(&device);
+        let ln = LayerNormConfig::new(d_model).init::<B>(&device);
+
+        let x = Tensor::<B, 3>::random([batch, 4, d_model], Distribution::Default, &device);
+        let xa = Tensor::<B, 3>::random([batch, 9, d_model * 2], Distribution::Default, &device);
+
+        layer_norm_cross_attn(&ln, &attn, x, xa);
+    }
 
     /// **The contract.** Stepping a sequence through the cache one token at a
     /// time must reproduce a single uncached pass over the whole thing.
@@ -306,11 +389,13 @@ mod tests {
             .context;
 
         // Cached: one token at a time. A lone query needs no mask.
-        let mut cache: Option<AttnKv<B>> = None;
+        let mut cache: Option<AttnKvPair<B>> = None;
         let mut steps = Vec::with_capacity(seq);
         for t in 0..seq {
             let step = x.clone().slice_dim(1, t as isize..(t + 1) as isize);
-            steps.push(layer_norm_self_attn_kv(&ln, &mha, step, None, &mut cache));
+            steps.push(layer_norm_self_attn_w_kv_cache(
+                &ln, &mha, step, None, &mut cache,
+            ));
         }
 
         assert_eq!(cache.as_ref().unwrap().seq_len(), seq);
@@ -334,8 +419,8 @@ mod tests {
 
         // Prefill 4, then step the last 2.
         let split = 4;
-        let mut cache: Option<AttnKv<B>> = None;
-        let prefill = layer_norm_self_attn_kv(
+        let mut cache: Option<AttnKvPair<B>> = None;
+        let prefill = layer_norm_self_attn_w_kv_cache(
             &ln,
             &mha,
             x.clone().slice_dim(1, 0..split as isize),
@@ -345,13 +430,15 @@ mod tests {
         let mut parts = vec![prefill];
         for t in split..seq {
             let step = x.clone().slice_dim(1, t as isize..(t + 1) as isize);
-            parts.push(layer_norm_self_attn_kv(&ln, &mha, step, None, &mut cache));
+            parts.push(layer_norm_self_attn_w_kv_cache(
+                &ln, &mha, step, None, &mut cache,
+            ));
         }
         let mixed: Tensor<B, 3> = Tensor::cat(parts, 1);
 
         // All in one go.
-        let mut cache2: Option<AttnKv<B>> = None;
-        let at_once = layer_norm_self_attn_kv(
+        let mut cache2: Option<AttnKvPair<B>> = None;
+        let at_once = layer_norm_self_attn_w_kv_cache(
             &ln,
             &mha,
             x,
@@ -383,10 +470,10 @@ mod tests {
             .forward(MhaInput::new(ln.forward(x.clone()), xa.clone(), xa.clone()))
             .context;
 
-        let kv = project_kv(&mha, xa);
+        let kv = project_kv_pair(&mha, xa);
         assert_eq!(kv.seq_len(), cross_len);
 
-        let cached = layer_norm_cross_attn_kv(&ln, &mha, x, &kv);
+        let cached = layer_norm_cross_attn_w_kv_cache(&ln, &mha, x, &kv);
 
         cached
             .to_data_as::<F>()
