+++ crates/bunsen/src/ops/transformers/attention/causal_mask.rs
@@ -1,3 +1,5 @@
+//! Causal mask utilities.
+
 use burn::{
     Tensor,
     prelude::{
