+++ crates/bunsen/src/ops/transformers/mod.rs
@@ -1,0 +1,2 @@
+//! Transformer operations.
+pub mod attention;
